chore: rename tab components
