+++ packages/components/tabs/src/index.ts
@@ -1,9 +1,9 @@
 export { OkuTabs } from './tabs'
-export { OkuTabList } from './tab-list'
-export { OkuTabTrigger } from './tab-trigger'
-export { OkuTabContent } from './tab-content'
+export { OkuTabList } from './tabList'
+export { OkuTabTrigger } from './tabTrigger'
+export { OkuTabContent } from './tabContent'
 
 export type { TabsProps } from './tabs'
-export { type TabListProps } from './tab-list'
-export { type TabsTriggerProps } from './tab-trigger'
-export { type TabsContentProps } from './tab-content'
+export { type TabListProps } from './tabList'
+export { type TabsTriggerProps } from './tabTrigger'
+export { type TabsContentProps } from './tabContent'
