+++ packages/components/tabs/src/tabTrigger.ts
